api-generation/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
PyJWT==2.3.0
passlib==1.7.4
bcrypt==3.2.0
slowapi==0.1.4
streamlit==1.22.0
requests==2.26.0